add a module for common fns used in predict
